sharemodal, screenshottable (iOS), integrate MarqueKit
