+++ Components/Globe/Views/Globe.AccountsView.swift
@@ -2,7 +2,7 @@
 //  Globe.AccountsView.swift
 //  Loom
 //
-//  Created by Ritesh Pakala on 8/18/23.
+//  Created by PEXAVC on 8/18/23.
 //
 
 import Foundation
