+++ Loom.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -10,6 +10,15 @@
       }
     },
     {
+      "identity" : "granite",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/pexavc/Granite.git",
+      "state" : {
+        "branch" : "main",
+        "revision" : "30a07a10f5fcb280b83a43fbb78761f7ab2eb782"
+      }
+    },
+    {
       "identity" : "highlightr",
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/raspu/Highlightr.git",
@@ -43,6 +52,15 @@
       "state" : {
         "branch" : "main",
         "revision" : "9a2b5ff0b4d7743659476958aed1569660093c90"
+      }
+    },
+    {
+      "identity" : "markdownview",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/pexavc/MarkdownView",
+      "state" : {
+        "branch" : "main",
+        "revision" : "541741eb819c4ce404390afb5120ba082c1e4187"
       }
     },
     {
