+++ Shared/Extensions/MarkdownFontGroup.swift
@@ -2,7 +2,7 @@
 //  MarkdownFontGroup.swift
 //  Loom (macOS)
 //
-//  Created by Ritesh Pakala on 8/19/23.
+//  Created by PEXAVC on 8/19/23.
 //
 
 import Foundation
