+++ Shared/Extensions/View+Size.swift
@@ -2,7 +2,7 @@
 //  View+Size.swift
 //  Loom
 //
-//  Created by Ritesh Pakala on 8/19/23.
+//  Created by PEXAVC on 8/19/23.
 //
 
 import Foundation
