+++ Shared/Utility/Snapshot/Snapshot.swift
@@ -2,7 +2,7 @@
 //  Snapshot.swift
 //  Loom (macOS)
 //
-//  Created by Ritesh Pakala on 8/19/23.
+//  Created by PEXAVC on 8/19/23.
 //
 
 import Foundation
@@ -10,99 +10,24 @@ import SwiftUI
 import Granite
 import GraniteUI
 import LemmyKit
+import MarqueKit
+
+#if os(iOS)
 import UIKit
-
-extension View {
-    func snapshot() -> UIImage {
-        
-        let controller = UIHostingController(rootView: self)
-        let view = controller.view
-
-        let targetSize = controller.view.intrinsicContentSize
-        view?.bounds = CGRect(origin: .zero, size: targetSize)
-        view?.backgroundColor = UIColor(Color.background)
-
-        let renderer = UIGraphicsImageRenderer(size: targetSize)
-
-        return renderer.image { _ in
-            view?.drawHierarchy(in: controller.view.bounds, afterScreenUpdates: true)
-        }
-    }
-}
-
-extension PostCardView {
-    static func snapshot(model: PostView, metadata: PageableMetadata?) -> UIImage {
-        PostCardView()
-            .contentContext(.init(postModel: model))
-            .environment(\.pagerMetadata, metadata)
-            .snapshot()
-    }
-}
-
-public struct ShareModal<Content: View>: View {
-    
-    @State var isScreenshotting: Bool = false
-    
-    var urlString: String? = nil
-    var content: () -> Content
-    init(urlString: String? = nil,
-         @ViewBuilder content: @escaping () -> Content) {
-        self.urlString = urlString
-        self.content = content
-    }
-    
-    public var body: some View {
-        VStack(spacing: 0) {
-            ScrollView {
-                ScreenshotView($isScreenshotting) {
-                    content()
-                }
-                .fixedSize(horizontal: false, vertical: true)
-                .cornerRadius(8)
-            }
-            
-            Spacer()
-            
-            HStack(spacing: .layer3) {
-                Spacer()
-                
-                Button {
-                    GraniteHaptic.light.invoke()
-                    isScreenshotting = true
-                } label: {
-                    Image(systemName: "photo")
-                        .font(.title2)
-                }
-                .buttonStyle(.plain)
-                
-                if let urlString {
-                    Button {
-                        GraniteHaptic.light.invoke()
-                        ModalService.share(urlString: urlString)
-                    } label: {
-                        Image(systemName: "link")
-                            .font(.title2)
-                            .scaleEffect(x: -1, y: 1)
-                    }
-                    .buttonStyle(.plain)
-                }
-                
-                Spacer()
-            }
-            
-            Spacer()
-        }
-    }
-}
-
+#else
+import AppKit
+#endif
 
 public struct ScreenshotView<Content: View> : GenericViewRepresentable {
     
     @Binding var isScreenshotting: Bool
+    var encodeMessage: String?
     var content: () -> Content
     public init(_ isScreenshotting: Binding<Bool>,
+                encodeMessage: String? = nil,
                 @ViewBuilder content: @escaping () -> Content) {
         self._isScreenshotting = isScreenshotting
+        self.encodeMessage = encodeMessage
         self.content = content
     }
     
@@ -125,21 +50,28 @@ public struct ScreenshotView<Content: View> : GenericViewRepresentable {
             uiView.drawHierarchy(in: uiView.bounds, afterScreenUpdates: true)
         }
         
-        ModalService
-            .share(image: image)
+        if let encodeMessage {
+            let result = MarqueKit.shared.encode(encodeMessage, withImage: image)
+            ModalService
+                .share(image: result.data)
+            
+            print(MarqueKit.shared.decode(image: result.data).payload)
+        } else {
+            ModalService
+                .share(image: image)
+        }
+        
     }
     #else
-    public func makeNSView(context: GenericControllerRepresentableContext<UIView>) -> SCNView {
+    public func makeNSView(context: Content) -> NSView {
         
-        scene.setup(data, force: data.isEmpty)
+        //TODO:
         
-        return scene.sceneView
+        return .init()
     }
 
-    public func updateNSView(_ scnView: SCNView, context: Context) {
-        if data.isNotEmpty {
-            scene.setup(data)
-        }
+    public func updateNSView(_ nsView: NSView, context: Context) {
+        //TODO: 
     }
     #endif
 }
