+++ Shared/Views/GraniteScrollView/GraniteScrollView+macOS.swift
@@ -2,7 +2,7 @@
 //  GraniteScrollView+Mac.swift
 //  Loom
 //
-//  Created by Ritesh Pakala on 8/18/23.
+//  Created by PEXAVC on 8/18/23.
 //
 #if os(macOS)
 import Foundation
