+++ Shared/Views/LogoView.swift
@@ -2,7 +2,7 @@
 //  LogoView.swift
 //  Loom
 //
-//  Created by Ritesh Pakala on 8/17/23.
+//  Created by PEXAVC on 8/17/23.
 //
 
 import Foundation
